Make the legacy integer sources inclusive and fix Int128Source
IntegerSource, LongSource, RandomNumberSource and their nullable variants
treated max as exclusive, so new IntegerSource(1, 3) never produced 3 while
NumberSource<T> and the date sources are inclusive on both ends. Int128Source
was worse: it built a value from two non-negative draws and clamped it to the
range, so the default source never produced a negative value and a restricted
range piled up on its bounds.

The uniform inclusive pick of NumberSourceBase moves to
Random.NextInclusive in Util, unchanged, and all four sources share it now.
NumberSource keeps its sequences, the long and Int128 sequences changed, the
integer ones did not (the rejection bound only differs when the top value is
drawn). All four throw ArgumentOutOfRangeException when max is below min.

Also: LongSourceBase.SetMinMax takes long instead of int, it could not reach
the range of its own source before, and Int128IdSource.SetStartValue takes
Int128 instead of long. Both are source compatible.

README gets a breaking changes block at the top of the 6.0 section.

845 tests per framework.

+++ holonsoft.AutoPoco/DataSources/Primitives/Int128IdSource.cs
@@ -10,7 +10,7 @@ public class Int128IdSource(Int128 startValue) : DataSourceBase<Int128> {
 
    public Int128IdSource() : this(Int128.Zero) { }
 
-   public Int128IdSource SetStartValue(long startValue) {
+   public Int128IdSource SetStartValue(Int128 startValue) {
       _currentId = startValue;
       return this;
    }
